pyproject.toml
[tool.poetry]
name = "mongo2pq"
version = "0.1.0"
description = ""
authors = ["Tadeas Uhlir <tadeas.uhlir@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
pymongo = "^4.5.0"
pyarrow = "^13.0.0"
motor = "^3.3.1"
tqdm = "^4.66.1"
pyyaml = "^6.0.1"


[tool.poetry.group.dev.dependencies]
ipython = "^8.16.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
